Atualizando.

+++ primeiro-exemplo/src/main/java/com/teste/primeiroexemplo/controller/ProdutoController.java
@@ -30,7 +30,7 @@ public class ProdutoController {
 	}
 	
 	@GetMapping("/{id}")
-	public Optional<Produto> obterPorId(@PathVariable Long id){
+	public Optional<Produto> obterPorId(@PathVariable Integer id){
 		return produtoService.obterPorId(id);
 	}
 	
@@ -40,13 +40,13 @@ public class ProdutoController {
 	}
 	
 	@DeleteMapping("/{id}")
-	public String deletar(@PathVariable Long id) {
+	public String deletar(@PathVariable Integer id) {
 		produtoService.deletar(id);
 		return "produto com o id: " + id + " foi deletado com sucesso!";
 	}
 	
 	@PutMapping("/{id}")
-	public Produto atualizar(@RequestBody Produto produto, @PathVariable Long id) {
+	public Produto atualizar(@RequestBody Produto produto, @PathVariable Integer id) {
 		return produtoService.atualizar(id, produto);
 	}
 }
